requirements.txt
Django==4.1.2 
django-extensions==3.2.1  
django-filter==22.1   
django-shell-plus==1.1.7  
djangorestframework==3.14.0 
Pygments==2.13.0 
parameterized==0.8.1
pytest==7.2.0
pytest-django==4.5.2
pyflakes==2.5.0
flake8==5.0.4